FIRST COMMIT
Setup of project, written Readme and carried out some planning, written a couple
of acceptance criteria and user stories. Feel like they are too large at the moment, I
am aware that i am supposed to be testing public interfaces of the code so mesesages
being sent to the API and responses from the request would be a good start, so perhaps
first test would be a request to the API to check that it is working!!? So would expect
a 200 response to a GET / HEAD request.