API: saving response as a variable.
Using postman app i found the 'login' field from the github api response, which appears to relate
to the user profile, so have used that to confirm that the response from the get request
is saved to the data variable. Believe that this is testing behaviour over state as it is checking
that the data has been saved, its checking the request for data and the process of it being saved.