STUB API REQUESTS
Took a while to do this the first time i tried, but having done it now it seems fairly
straight forward. Simply i have downloaded the API responses to a './spec/support' folder
that has data for the two test cases i am trying, the webmock get disables net connections
therefore the data comes form (when config.before) the spec_helper / this json file,
which is parsed by the program which now gives us data on file, not remote and stubbed with
accurate copy of what was available. I think this is pretty good, clearly web requests are
currently stubbed to these two repositories, but it will work when sending a request to any
others.... will add a note to the readme. I am concious that i am only working from the master branch of the project