REFACTOR API requests
I had a nagging doubt about the get requests, so instead of using real world repo names
have converted them to TEST1 and TEST2 in the stub and file names. Hopefully it will make
it clearer that it is a stub / test data and not a real request. Clarity is important.