PRINT_OUT : big update to the printout method
So the acceptance criteria are now passing following some private methods. Sandi Metz says that we are
not supposed to test private methods so i have not, the print_out method now has some new features,
to accomodate the test... that was quite a step, but I believe it works at the moment.

Committing now as has passed the test, didnt commit earlier as it wasnt. The printout hopefully is self
explanatory, the create language list does its namesake and collects an array of the languages used in the
@data, i.e. getting the info we need to analyse. Then most common does its name sake and checks the language
list and finds the most common language that occurs, which sets the result.. i.e. the most common and therefore
assumed favourite language.

print results does its namesake and prints the results in a nice string, with the @result to pass the test.

Hopefully that makes sense! Acceptance criteria passing, now to apply some more tests and refactor a bit, feeling
like i have done it all in one file and I could create a super class that requests API info, this being a git hub api
request which is a child of that, or maybe a module for the request / printer... but they are quite tied to the
data in the class. Need to have a think.

I could also make the print_out method private and call it at the bottom of the request method, but I think it may be
nice for the user to do it seperatly... maybe. Would be less public methods if it was in private and called at bottom of
the request. Main thought at the moment is to apply more tests as have got it done without many tests and feel like
there should be more to check edge cases.