REFACTOR: request method
Nice bit of refactoring this i think, felt like encapsulting the requester method
away from the favourite language as they are seperate verbs and so probably
should be seperate objects.

To save the existing tests ive created a GitHubApiDataRequester object that
is default injected into the favourite language class, but clearly it can be
replaced with a request to another api or resource, with this design you can
also replace the input to the request so the request to other resources can
do exactly this. Nice little refactor i think.