REFACTOR some work on reusability.
Have added some notes on thoughts, as ive completed the requirement (i hope!!), but the
code can do with some cleaning up to make it more solid, i.e. i dont think that this
essentially needs to be a favourite language request, it can be just a favourite request
and the params can define what key is being searched for i.e. the print statment can take
params from the user and define what they want to loop through the data set for. Interesting
stuff! Naming is becoming a little harder as need params to be clear and the extraction is
making it less clear, have extracted to a level that i am comfortable with the naming
at the moment. Next step may be to change the name of the class itself, from
FavouriteLanguage to Favourite... nice..