REFACTOR: tidy up / named parameters
added named parameters on the request, removed the parsing of a json from
the github api request - seemed better for single responsability, and maily
removed the language element from all the file names. It is now not reliant
on language and works to test your favourite whatever from whatever data set.